Fixed office bug

+++ capstone/java/database/schema.sql
@@ -20,7 +20,7 @@ CREATE TABLE office_info (
     phone_number bigint NOT NULL,
     office_hours_open time NOT NULL,
     office_hours_close time NOT NULL,
-    cost_per_hour money NOT NULL,
+    cost_per_hour int NOT NULL,
     office_image_url varchar(500),
     delay boolean NOT NULL,
     CONSTRAINT PK_office_info PRIMARY KEY (office_id)
